更新依赖

+++ pkg/bot/proto2mirai.go
@@ -139,7 +139,7 @@ func ProtoFaceToMiraiFace(data map[string]string) message.IMessageElement {
 		return EmptyText()
 	}
 	return &message.FaceElement{
-		FaceID: uint16(id),
+		FaceID: uint32(id),
 	}
 }
